fix build break on linux
